Auto table creation with JPA

+++ src/main/java/io/github/enzopavani/libraryapi/model/Livro.java
@@ -4,13 +4,14 @@ import io.github.enzopavani.libraryapi.model.enums.GeneroLivro;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
 @Entity
 @Table(name="livro")
 /*
-    - Anotação @Data inclui
+- Anotação @Data inclui
 @Getter@Setter
 @ToString
 @EqualsAndHashCode
@@ -38,7 +39,7 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name="preco", precision=18, scale=2)
-    private Double preco;
+    private BigDecimal preco;
 
     @ManyToOne
     @JoinColumn(name="id_autor")
